+++ genomepy/cli.py
@@ -48,10 +48,16 @@ def genomes(provider=None):
 
 # extended options for genomepy install
 general_install_options = {
+    "provider": {
+        "short": "p",
+        "long": "provider",
+        "help": "download from this provider",
+        "default": None,
+    },
     "genomes_dir": {
         "short": "g",
         "long": "genomes_dir",
-        "help": "genomes directory",
+        "help": "create output directory here",
         "default": None,
     },
     "localname": {
@@ -178,7 +184,6 @@ def custom_options(options):
 
 
 @custom_options(get_install_options())
-@click.argument("provider")
 @click.argument("name")
 @cli.command()
 def install(
@@ -197,10 +202,10 @@ def install(
     force,
     **kwargs,
 ):
-    """Install genome NAME from provider PROVIDER in directory GENOME_DIR."""
+    """install a genome & run active plugins"""
     genomepy.install_genome(
         name,
-        provider,
+        provider=provider,
         genomes_dir=genomes_dir,
         localname=localname,
         mask=mask,